refactor: split showScore
