+++ src/pages/SnakeGamePage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import GamePage from "../components/snakeGame";
+import GamePage from "../components/snakegame/snakeGame";
 
 export const SnakeGame = () => {
   return <GamePage />;
