+++ src/types/ShowScoreProps.d.ts
@@ -1,0 +1,8 @@
+import { SnakeGameType } from './SnakeGameType';
+import { useState } from 'react';
+
+export interface ShowScoreProps {
+    snake : SnakeGameType[];
+    score : number;
+    restProps: any;
+}
